% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_spectrum-methods.R
\name{c.libs_spectrum}
\alias{c.libs_spectrum}
\title{Combine Multiple LIBS Spectra by Stacking Shots}
\usage{
\method{c}{libs_spectrum}(...)
}
\arguments{
\item{...}{Two or more \code{libs_spectrum} objects.}
}
\value{
A new \code{libs_spectrum} with combined shots.
}
\description{
Concatenates shots from multiple spectra. All spectra must share the same
wavelength axis.
}
