% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_dataset-methods.R
\name{dim.libs_dataset}
\alias{dim.libs_dataset}
\title{Dimensions of a Dataset (spectra x channels)}
\usage{
\method{dim}{libs_dataset}(x)
}
\arguments{
\item{x}{A \code{libs_dataset} object.}
}
\value{
Integer vector of length 2.
}
\description{
Dimensions of a Dataset (spectra x channels)
}
