% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_spectrum-methods.R
\name{dim.libs_spectrum}
\alias{dim.libs_spectrum}
\title{Dimensions of a LIBS Spectrum (shots x channels)}
\usage{
\method{dim}{libs_spectrum}(x)
}
\arguments{
\item{x}{A \code{libs_spectrum} object.}
}
\value{
Integer vector of length 2.
}
\description{
Dimensions of a LIBS Spectrum (shots x channels)
}
