% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_calibration-class.R
\name{is_libs_calibration}
\alias{is_libs_calibration}
\title{Test whether an object is a libs_calibration}
\usage{
is_libs_calibration(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a libs_calibration
}
