% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_dataset-class.R
\name{is_libs_dataset}
\alias{is_libs_dataset}
\title{Test whether an object is a libs_dataset}
\usage{
is_libs_dataset(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a libs_dataset
}
