% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_spectrum-class.R
\name{is_libs_spectrum}
\alias{is_libs_spectrum}
\title{Test whether an object is a libs_spectrum}
\usage{
is_libs_spectrum(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a libs_spectrum
}
