% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_dataset-methods.R
\name{length.libs_dataset}
\alias{length.libs_dataset}
\title{Number of Spectra in a Dataset}
\usage{
\method{length}{libs_dataset}(x)
}
\arguments{
\item{x}{A \code{libs_dataset} object.}
}
\value{
Integer scalar.
}
\description{
Number of Spectra in a Dataset
}
