% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_spectrum-methods.R
\name{length.libs_spectrum}
\alias{length.libs_spectrum}
\title{Number of Channels in a LIBS Spectrum}
\usage{
\method{length}{libs_spectrum}(x)
}
\arguments{
\item{x}{A \code{libs_spectrum} object.}
}
\value{
Integer scalar.
}
\description{
Number of Channels in a LIBS Spectrum
}
