% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libscanR-package.R
\docType{package}
\name{libscanR-package}
\alias{libscanR}
\alias{libscanR-package}
\title{libscanR: Analysis and Visualization of LIBS Spectra}
\description{
A vendor-agnostic pipeline for importing, preprocessing, calibrating, and
visualizing Laser-Induced Breakdown Spectroscopy (LIBS) spectral data with
a focus on biomedical tissue analysis.
}
\section{Main functions}{

\itemize{
\item Import: \code{\link[=ls_read_spectrum]{ls_read_spectrum()}}, \code{\link[=ls_read_dir]{ls_read_dir()}}, \code{\link[=ls_read_auto]{ls_read_auto()}}
\item Preprocessing: \code{\link[=ls_baseline]{ls_baseline()}}, \code{\link[=ls_normalize]{ls_normalize()}}, \code{\link[=ls_smooth]{ls_smooth()}}, \code{\link[=ls_crop]{ls_crop()}}
\item Peak analysis: \code{\link[=ls_find_peaks]{ls_find_peaks()}}, \code{\link[=ls_identify_peaks]{ls_identify_peaks()}}, \code{\link[=ls_element_db]{ls_element_db()}}
\item Calibration: \code{\link[=ls_calibrate]{ls_calibrate()}}, \code{\link[=ls_saha_boltzmann]{ls_saha_boltzmann()}}, \code{\link[=ls_quantify]{ls_quantify()}}
\item Chemometrics: \code{\link[=ls_pca]{ls_pca()}}, \code{\link[=ls_plsda]{ls_plsda()}}, \code{\link[=ls_cluster]{ls_cluster()}}
\item Spatial mapping: \code{\link[=ls_build_map]{ls_build_map()}}, \code{\link[=ls_map_elements]{ls_map_elements()}}
\item Tissue analysis: \code{\link[=ls_tissue_classify]{ls_tissue_classify()}}, \code{\link[=ls_tissue_discriminate]{ls_tissue_discriminate()}}
\item Interactive app: \code{\link[=ls_run_app]{ls_run_app()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-heller/libscanR}
  \item \url{https://r-heller.github.io/libscanR/}
  \item Report bugs at \url{https://github.com/r-heller/libscanR/issues}
}

}
\author{
\strong{Maintainer}: Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

}
\keyword{internal}
