% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{ls_loq}
\alias{ls_loq}
\title{Limit of Quantification (10-sigma)}
\usage{
ls_loq(calibration, blank = NULL, window_nm = 1)
}
\arguments{
\item{calibration}{A \code{\link[=ls_calibration]{ls_calibration()}} object.}

\item{blank}{Optional \code{\link[=ls_spectrum]{ls_spectrum()}} of a blank sample.}

\item{window_nm}{Numeric. Integration window. Default 1.}
}
\value{
Numeric LOQ in the calibration's concentration unit.
}
\description{
Calculates LOQ using the 10-sigma criterion.
}
\examples{
ds <- ls_example_data("calibration")
conc <- ds$sample_info$concentration
cal <- ls_calibrate(ds, "Ca", 393.37, conc, verbose = FALSE)
ls_loq(cal)
}
