% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{ls_plot_element_map}
\alias{ls_plot_element_map}
\title{Plot an Element Map (alias)}
\usage{
ls_plot_element_map(map, ...)
}
\arguments{
\item{map}{A \code{libs_map} object.}

\item{...}{Passed to \code{\link[=ls_plot_map]{ls_plot_map()}}.}
}
\value{
A ggplot2 object.
}
\description{
Convenience alias for \code{\link[=ls_plot_map]{ls_plot_map()}}.
}
\examples{
ds <- ls_example_data("spatial")
ls_plot_element_map(ls_build_map(ds, "Ca", 393.37))
}
