% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{ls_plot_map}
\alias{ls_plot_map}
\title{Plot an Elemental Map}
\usage{
ls_plot_map(map, color_scale = "viridis", title = NULL)
}
\arguments{
\item{map}{A \code{libs_map} object from \code{\link[=ls_build_map]{ls_build_map()}}.}

\item{color_scale}{Character. \code{"viridis"}, \code{"magma"}, \code{"plasma"},
\code{"inferno"}, \code{"hot"}, \code{"jet"}. Default \code{"viridis"}.}

\item{title}{Character. Plot title. Default: element symbol.}
}
\value{
A ggplot2 object.
}
\description{
Plot an Elemental Map
}
\examples{
ds <- ls_example_data("spatial")
m <- ls_build_map(ds, "Ca", 393.37)
ls_plot_map(m)
}
