% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{ls_plot_map_panel}
\alias{ls_plot_map_panel}
\title{Plot Multi-Element Map Panel}
\usage{
ls_plot_map_panel(maps, ncol = 3)
}
\arguments{
\item{maps}{Named list of \code{libs_map} objects.}

\item{ncol}{Integer. Facet columns. Default 3.}
}
\value{
A ggplot2 object.
}
\description{
Plots multiple elemental maps as a faceted panel.
}
\examples{
ds <- ls_example_data("spatial")
ms <- ls_map_elements(ds, c("Ca", "Fe", "Zn"),
                      c(Ca = 393.37, Fe = 371.99, Zn = 213.86))
ls_plot_map_panel(ms)
}
