% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-chemometrics.R
\name{ls_plot_pca}
\alias{ls_plot_pca}
\title{Plot PCA Scores}
\usage{
ls_plot_pca(pca, pc_x = 1, pc_y = 2, color_by = NULL, ellipses = TRUE)
}
\arguments{
\item{pca}{A \code{libs_pca} object from \code{\link[=ls_pca]{ls_pca()}}.}

\item{pc_x}{Integer. PC for x-axis. Default 1.}

\item{pc_y}{Integer. PC for y-axis. Default 2.}

\item{color_by}{Character. Grouping variable. Default \code{NULL}.}

\item{ellipses}{Logical. Draw 95\% confidence ellipses per group. Default \code{TRUE}.}
}
\value{
A ggplot2 object.
}
\description{
Plot PCA Scores
}
\examples{
ds <- ls_example_data("tissue")
pca <- ls_pca(ds, n_components = 4)
ls_plot_pca(pca, color_by = "tissue")
}
