% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-calibration.R
\name{ls_plot_residuals}
\alias{ls_plot_residuals}
\title{Plot Calibration Residuals}
\usage{
ls_plot_residuals(calibration)
}
\arguments{
\item{calibration}{A \code{\link[=ls_calibration]{ls_calibration()}} object.}
}
\value{
A ggplot2 object.
}
\description{
Standard regression diagnostic: residuals vs fitted.
}
\examples{
ds <- ls_example_data("calibration")
conc <- ds$sample_info$concentration
cal <- ls_calibrate(ds, "Ca", 393.37, conc, verbose = FALSE)
ls_plot_residuals(cal)
}
