% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-chemometrics.R
\name{ls_plot_scree}
\alias{ls_plot_scree}
\title{Plot PCA Variance (Scree Plot)}
\usage{
ls_plot_scree(pca)
}
\arguments{
\item{pca}{A \code{libs_pca} object.}
}
\value{
A ggplot2 object.
}
\description{
Plot PCA Variance (Scree Plot)
}
\examples{
ds <- ls_example_data("tissue")
pca <- ls_pca(ds, n_components = 6)
ls_plot_scree(pca)
}
