% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{ls_quantify}
\alias{ls_quantify}
\title{Quantify Element Concentration}
\usage{
ls_quantify(calibration, x, window_nm = NULL)
}
\arguments{
\item{calibration}{A \code{\link[=ls_calibration]{ls_calibration()}} object.}

\item{x}{A \code{\link[=ls_spectrum]{ls_spectrum()}} or \code{\link[=ls_dataset]{ls_dataset()}} object.}

\item{window_nm}{Numeric. Peak integration window. Defaults to 1 if NULL.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{sample_id}, \code{element},
\code{concentration}, \code{unit}, \code{below_lod}, \code{below_loq}.
}
\description{
Applies a calibration model to unknown spectra.
}
\examples{
ds <- ls_example_data("calibration")
conc <- ds$sample_info$concentration
cal <- ls_calibrate(ds, "Ca", 393.37, conc, verbose = FALSE)
ls_quantify(cal, ds)
}
