% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-read.R
\name{ls_read_csv}
\alias{ls_read_csv}
\title{Read a LIBS Spectrum from CSV (Convenience Wrapper)}
\usage{
ls_read_csv(path, ...)
}
\arguments{
\item{path}{Character. Path to CSV file.}

\item{...}{Additional arguments passed to \code{\link[=ls_read_spectrum]{ls_read_spectrum()}}.}
}
\value{
A \code{libs_spectrum} object.
}
\description{
Read a LIBS Spectrum from CSV (Convenience Wrapper)
}
