% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_dataset-methods.R
\name{plot.libs_dataset}
\alias{plot.libs_dataset}
\title{Plot a LIBS Dataset}
\usage{
\method{plot}{libs_dataset}(x, y, ...)
}
\arguments{
\item{x}{A \code{libs_dataset} object.}

\item{y}{Ignored.}

\item{...}{Passed to \code{\link[=ls_plot_overlay]{ls_plot_overlay()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a LIBS Dataset
}
