% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_spectrum-methods.R
\name{plot.libs_spectrum}
\alias{plot.libs_spectrum}
\title{Plot a LIBS Spectrum}
\usage{
\method{plot}{libs_spectrum}(x, y, ...)
}
\arguments{
\item{x}{A \code{libs_spectrum} object.}

\item{y}{Ignored.}

\item{...}{Arguments passed to \code{\link[=ls_plot_spectrum]{ls_plot_spectrum()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a LIBS Spectrum
}
