% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_calibration-class.R
\name{predict.libs_calibration}
\alias{predict.libs_calibration}
\title{Predict Concentrations from a Calibration Model}
\usage{
\method{predict}{libs_calibration}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{libs_calibration} object.}

\item{newdata}{A numeric vector of new intensities (or a matrix for PLS).}

\item{...}{Unused.}
}
\value{
Numeric vector of predicted concentrations.
}
\description{
Predict Concentrations from a Calibration Model
}
