% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_calibration-class.R
\name{print.libs_calibration}
\alias{print.libs_calibration}
\title{Print a LIBS Calibration}
\usage{
\method{print}{libs_calibration}(x, ...)
}
\arguments{
\item{x}{A \code{libs_calibration} object.}

\item{...}{Unused.}
}
\value{
Invisibly \code{x}.
}
\description{
Print a LIBS Calibration
}
