% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_dataset-methods.R
\name{print.libs_dataset}
\alias{print.libs_dataset}
\title{Print a LIBS Dataset}
\usage{
\method{print}{libs_dataset}(x, ...)
}
\arguments{
\item{x}{A \code{libs_dataset} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a LIBS Dataset
}
