% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_spectrum-methods.R
\name{print.libs_spectrum}
\alias{print.libs_spectrum}
\title{Print a LIBS Spectrum}
\usage{
\method{print}{libs_spectrum}(x, ...)
}
\arguments{
\item{x}{A \code{libs_spectrum} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a LIBS Spectrum
}
