% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_dataset-methods.R
\name{[.libs_dataset}
\alias{[.libs_dataset}
\title{Subset a LIBS Dataset}
\usage{
\method{[}{libs_dataset}(x, i)
}
\arguments{
\item{x}{A \code{libs_dataset} object.}

\item{i}{Integer or logical index of spectra to keep.}
}
\value{
A new \code{libs_dataset}.
}
\description{
Subset a LIBS Dataset
}
