% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_calibration-class.R
\name{summary.libs_calibration}
\alias{summary.libs_calibration}
\title{Summary of a LIBS Calibration}
\usage{
\method{summary}{libs_calibration}(object, ...)
}
\arguments{
\item{object}{A \code{libs_calibration} object.}

\item{...}{Unused.}
}
\value{
Invisibly a list of key figures of merit.
}
\description{
Summary of a LIBS Calibration
}
