% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_dataset-methods.R
\name{summary.libs_dataset}
\alias{summary.libs_dataset}
\title{Summary of a LIBS Dataset}
\usage{
\method{summary}{libs_dataset}(object, ...)
}
\arguments{
\item{object}{A \code{libs_dataset} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns a list with per-group statistics.
}
\description{
Summary of a LIBS Dataset
}
