% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_spectrum-methods.R
\name{summary.libs_spectrum}
\alias{summary.libs_spectrum}
\title{Summary of a LIBS Spectrum}
\usage{
\method{summary}{libs_spectrum}(object, n_peaks = 5, ...)
}
\arguments{
\item{object}{A \code{libs_spectrum} object.}

\item{n_peaks}{Integer. Number of top peaks to show. Default 5.}

\item{...}{Unused.}
}
\value{
Invisibly returns a list with summary statistics.
}
\description{
Summary of a LIBS Spectrum
}
